docs: 사용되지 않는 테스트 파일 정리

+++ src/admin/admin.repository.ts
@@ -15,71 +15,40 @@ export class AdminRepository {
     private readonly commentRepository: Repository<CommentEntity>,
   ) {}
 
-  async checkUser(
-    data: kakaoPayload,
-  ): Promise<{ userId: string; nickname: string }> {
+  async checkUser(data: kakaoPayload): Promise<{ userId: string; nickname: string }> {
     const { socialId, social } = data;
-    const user = await this.usersRepository
-      .createQueryBuilder()
-      .select(['USER.userId', 'USER.nickname'])
-      .from(UserEntity, 'USER')
-      .where('USER.socialId = :socialId', { socialId })
-      .andWhere('USER.social = :social', { social })
-      .getOne();
+    const user = await this.usersRepository.createQueryBuilder().select(['USER.userId', 'USER.nickname']).from(UserEntity, 'USER').where('USER.socialId = :socialId', { socialId }).andWhere('USER.social = :social', { social }).getOne();
     return user;
   }
 
-  async createUser(
-    data: kakaoPayload,
-  ): Promise<{ userId: string; nickname: string }> {
+  async createUser(data: kakaoPayload): Promise<{ userId: string; nickname: string }> {
     const { socialId, social } = data;
     let newUser;
-    await this.usersRepository.manager.transaction(
-      async (transactionalEntityManager) => {
-        await this.usersRepository
-          .createQueryBuilder()
-          .insert()
-          .into(UserEntity)
-          .values({
-            ...data,
-          })
-          .execute();
-        newUser = await transactionalEntityManager
-          .createQueryBuilder()
-          .select(['USER.userId', 'USER.nickname'])
-          .from(UserEntity, 'USER')
-          .where('USER.socialId = :socialId', { socialId })
-          .andWhere('USER.social = :social', { social })
-          .getOne();
-      },
-    );
+    await this.usersRepository.manager.transaction(async (transactionalEntityManager) => {
+      await this.usersRepository
+        .createQueryBuilder()
+        .insert()
+        .into(UserEntity)
+        .values({
+          ...data,
+        })
+        .execute();
+      newUser = await transactionalEntityManager.createQueryBuilder().select(['USER.userId', 'USER.nickname']).from(UserEntity, 'USER').where('USER.socialId = :socialId', { socialId }).andWhere('USER.social = :social', { social }).getOne();
+    });
     return newUser;
   }
 
   async findById(userId: string) {
-    return await this.usersRepository
-      .createQueryBuilder('user')
-      .select()
-      .where('user.userId = :userId', { userId })
-      .getOne();
+    return await this.usersRepository.createQueryBuilder('user').select().where('user.userId = :userId', { userId }).getOne();
   }
 
   async deleteUser(userId: string) {
-    await this.usersRepository
-      .createQueryBuilder()
-      .delete()
-      .from(UserEntity)
-      .where('userId = :userId', { userId })
-      .execute();
+    await this.usersRepository.createQueryBuilder().delete().from(UserEntity).where('userId = :userId', { userId }).execute();
     return;
   }
 
   async findCommentList(userId: string): Promise<{ id: string }[]> {
-    return await this.commentRepository
-      .createQueryBuilder('comment')
-      .select(['comment.id'])
-      .where('comment.userId = :userId', { userId })
-      .getMany();
+    return await this.commentRepository.createQueryBuilder('comment').select(['comment.id']).where('comment.userId = :userId', { userId }).getMany();
   }
 
   async findCommentOptions(id: string): Promise<{
@@ -88,18 +57,6 @@ export class AdminRepository {
     summonerName: { summonerName: string };
     champId: { id: string };
   }> {
-    return await this.commentRepository
-      .createQueryBuilder('comment')
-      .leftJoinAndSelect('comment.champId', 'champ')
-      .leftJoinAndSelect('comment.summonerName', 'summoner')
-      .select([
-        'comment.category',
-        'comment.champId',
-        'comment.summonerName',
-        'champ.id',
-        'summoner.summonerName',
-      ])
-      .where('comment.id = :id', { id })
-      .getOne();
+    return await this.commentRepository.createQueryBuilder('comment').leftJoinAndSelect('comment.champId', 'champ').leftJoinAndSelect('comment.summonerName', 'summoner').select(['comment.category', 'comment.champId', 'comment.summonerName', 'champ.id', 'summoner.summonerName']).where('comment.id = :id', { id }).getOne();
   }
 }
